third_party/WebKit/LayoutTests/inspector/console/console-revoke-error-in-worker-expected.txt

CONSOLE ERROR: line 1: Uncaught (in promise) Error: Handled error
CONSOLE MESSAGE: Handler added to rejected promise
Tests that console revokes lazily handled promise rejections.

Creating worker with promise

Message added: error log
    1 error
    0 handled promise rejections
    0 warnings
console-message-wrapper console-error-level

Handling promise
console-message-wrapper console-revokedError-level
